Renamed xmonad start script (don't overwrite the command xmonad) Handle links requiring root access. Move startup stuff from xmonad.hs to xmonadStart

+++ makeLinks.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
 
 class Link(object):
-  def __init__(self,src,dest=None,computerSpecific=False):
+  def __init__(self,src,dest=None,computerSpecific=False,sudo=False):
     self.src = src
     self.dest = dest
     self.computerSpecific = computerSpecific
+    self.sudo = sudo
     if self.dest is None:
       self.dest = '.' + src
 
@@ -20,36 +21,52 @@ links = [
   Link('bin','bin'),
   Link('toprc'),
   Link('inputrc'),
-  Link('xmobarrc',computerSpecific=True)
+  Link('xmobarrc',computerSpecific=True),
+  Link('xmonad.desktop','/usr/share/xsessions/xmonad.desktop',sudo=True)
 ]
 
-import os, socket
+import os, socket, sys, subprocess
 
 def getLinkName(linkName):
   return os.path.expanduser(os.path.join('~/',linkName))
 
-def handleLink(linkName,source,make):
+def handleLink(linkName,source,make,sudo):
   link = getLinkName(linkName)
+  source = os.path.abspath(source)
   if make:
     try:
-      os.symlink(os.path.abspath(source),link)
-    except:
-      import sys
+      if sudo:
+        subprocess.check_call(['sudo','ln','-s',source,link])
+      else:
+        os.symlink(source,link)
+    except OSError as err:
       print >>sys.stderr,'Failed to make link for',source,link
+      print >>sys.stderr,err
   else:
     try:
-      os.unlink(link)
-    except OSError:
-      pass
+      if sudo:
+        subprocess.check_call(['sudo','unlink',link])
+      else:
+        os.unlink(link)
+    except OSError as err:
+      if err.errno != 2: # errno 2 is if no such file exists
+        print >>sys.stderr,err
 
-def handleLinks(make=True):
+def handleLinks(make=True,sudo=False):
   hostname = socket.gethostname()
   for link in links:
     src = link.src
     if link.computerSpecific:
       src += '-' + hostname
-    handleLink(link.dest,src,make)
+    if not(sudo) and link.sudo:
+      continue
+    handleLink(link.dest,src,make,link.sudo)
 
 if __name__ == '__main__':
-  handleLinks(make=False)
-  handleLinks(make=True)
+  from optparse import OptionParser
+  parser = OptionParser('makeLinks.py [options]')
+  parser.add_option('--sudo',action='store_true',dest='sudo',default=False,help='Do actions requiring sudo access')
+  options,args = parser.parse_args()
+  assert(len(args) == 0)
+  handleLinks(make=False,sudo=options.sudo)
+  handleLinks(make=True,sudo=options.sudo)
